This commit was manufactured by cvs2svn to create tag 'R-0-63-3'.
git-svn-id: https://svn.r-project.org/R/tags/R-0-63-3@3784 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/unix/devX11.c
@@ -1021,7 +1021,7 @@ static void X11_Circle(double x, double y, int coords,
 static void X11_Line(double x1, double y1, double x2, double y2,
 		     int coords, DevDesc *dd)
 {
-    double xx1, yy1, xx2, yy2;
+    int xx1, yy1, xx2, yy2;
     x11Desc *xd = (x11Desc *) dd->deviceSpecific;
 
     /* In-place conversion ok */
